M20 phase 4: 4-level paging (x86_64) + vmm.h uintptr_t widening
Lands the x86_64 implementation of kernel/includes/vmm.h on top of
the page tables boot.s built in Phase 2.  The API surface widens to
uintptr_t (4 bytes on i386, 8 bytes on x86_64) so the same prototype
serves both archs without #ifdef at call sites.

vmm.h
- vmm_map / vmm_map_4mib / vmm_unmap / vmm_translate /
  vmm_kernel_pd_phys all switch their virt/phys/return types from
  uint32_t to uintptr_t.  On i386 uintptr_t == uint32_t so no source
  change is needed at any call site.  Doc updated to explain the
  per-arch granule difference (4 MiB PSE on i386 vs. paired 2 MiB
  large pages on x86_64) for vmm_map_4mib.

i386 vmm.c
- File moved from kernel/mem/vmm.c -> kernel/hal/x86/vmm.c (it's
  i386-specific 2-level paging; the location now reflects that).
- Wrapper functions take uintptr_t, cast internally to uint32_t.
  Identity casts on i386, but the explicit cast makes the arch
  boundary obvious.
- vmm_kernel_pd_phys widened return; smp.c assignment still works
  via implicit truncation (i386 PD lives in the low 4 GiB).
- Makefile ARCH_C_SRCS for i386 updated to the new path.

x86_64 vmm.c (new)
- Inherits boot.s's PML4 / PDPT / PD via extern symbols rather than
  rebuilding the page tables from scratch — avoids the
  chicken-and-egg of touching kernel page tables while running on
  them, and reuses the first-1-GiB identity map for free.
- walk_to_pt() walks the 4-level hierarchy (PML4 -> PDPT -> PD ->
  PT), allocating intermediate tables on demand from pmm_alloc_frame.
  Refuses to refine an entry that's already a large-page (PS=1)
  mapping; same policy as i386's 4 MiB PSE refusal.
- vmm_map: standard 4 KiB install.
- vmm_map_4mib: installs two adjacent 2 MiB large PD entries, so
  the API contract of "map 4 MiB" holds end-to-end with i386
  callers unchanged.
- vmm_unmap / vmm_translate: walk and clear / return.
- vmm_print_status: dumps CR0/CR3/CR4 + EFER (with the LMA bit
  decoded so a serial log proves we're still in long mode).
- rdmsr cannot use the "=A" GCC asm constraint in 64-bit mode (it
  means "edx:eax pair as one 64-bit value", which is the legacy
  i386 split — long-mode rdmsr zero-extends into rax/rdx).  Use
  two separate =a / =d outputs and combine in C.

Build state
- i386 baseline rebuilds + boots cleanly (all M19/M18.5 self-tests
  PASS, shell prompt up).  Verified that the vmm.c relocation
  produced no observable behaviour change.
- x86_64 vmm.o compiles cleanly with -Wall -Wextra.  Link still
  fails on kprintf / pmm_alloc_frame / task_finish_first_switch /
  task_exit — all from kernel core that Phase 5 will add to
  CORE_C_SRCS for x86_64.

M20-PROGRESS.md
- Phase 4 status flipped to shipped.
- Lessons learned: API widening pattern (uintptr_t at boundary,
  arch-internal cast), inheriting boot.s page tables vs.
  rebuilding, vmm_map_4mib semantic preservation via paired 2 MiB
  pages, rdmsr "=A" constraint trap.
- Next concrete action updated to Phase 5 (link kernel_main).

+++ kernel/includes/vmm.h
@@ -1,20 +1,31 @@
 /* =============================================================================
- * vmm.h — Virtual Memory Manager public interface.
+ * vmm.h — Virtual Memory Manager public interface (arch-neutral API).
  *
- * After `vmm_init`, paging is on and the first 256 MiB of physical memory
- * is identity-mapped via 4 MiB PSE pages (base = virt, no translation).
- * That covers the kernel image, low memory, and any PMM-allocated frame
- * we currently care about, so every pointer we already hold keeps
- * working as-is.
+ * After `vmm_init`, paging is on and a sizeable chunk of physical
+ * memory is identity-mapped via the largest page granule the arch
+ * supports cheaply:
+ *
+ *   i386:   first 256 MiB via 4 MiB PSE pages.
+ *   x86_64: first 1 GiB via 2 MiB pages (set up in boot.s before C
+ *           runs; vmm.c inherits them).
+ *
+ * That covers the kernel image, low memory, and any PMM-allocated
+ * frame we currently care about, so every pointer we already hold
+ * keeps working as-is.
  *
  * `vmm_map` installs a 4 KiB-granular virtual→physical mapping using
- * conventional two-level tables, allocating a new page table from the
- * PMM when the relevant PDE is absent.  It explicitly refuses to modify
- * PDEs that are currently 4 MiB PSE entries (the kernel identity map)
- * because unpacking a 4 MiB mapping into 1024 × 4 KiB entries is
- * significantly more subtle and not needed yet.
+ * conventional N-level tables, allocating new intermediate tables
+ * from the PMM when the relevant entry is absent.  Like the i386
+ * impl, it refuses to refine an entry that is already a "large page"
+ * (PSE on i386, 2 MiB / 1 GiB on x86_64) — splitting a large page
+ * into 4 KiB ones is significantly more subtle and not needed yet.
  *
- * Paging reference: Intel SDM Vol 3, Chapter 4 ("Paging").
+ * Addresses are `uintptr_t` (4 bytes on i386, 8 bytes on x86_64) so
+ * the same prototype works on both arches without conditional
+ * compilation at call sites.  i386 callers see no source change;
+ * x86_64 callers can pass full 64-bit addresses.
+ *
+ * Paging reference: Intel SDM Vol 3 §4 (Paging), AMD64 APM Vol 2 §5.
  * ============================================================================= */
 
 #ifndef VMM_H
@@ -23,38 +34,51 @@
 #include <stdint.h>
 
 /* Flags passed to `vmm_map`.  Present bit is implicit — unmap if you want
- * a P=0 entry.  These constants match the hardware bit positions so they
- * can flow straight into PTEs. */
+ * a P=0 entry.  These constants match the hardware bit positions in BOTH
+ * i386 PTEs and x86_64 PTEs — Intel kept the low 12 bits compatible when
+ * adding long mode, so the same values flow straight into either page
+ * table format. */
 #define VMM_WRITABLE     0x002
 #define VMM_USER         0x004
 #define VMM_WRITE_THRU   0x008
 #define VMM_CACHE_DIS    0x010
 
-/* Turn on paging + identity-map the kernel region.  Must be called after
- * `pmm_init` so subsequent `vmm_map` calls can allocate page tables. */
+/* Turn on paging + identity-map the kernel region.  On i386 this builds
+ * the page directory and turns on CR0.PG; on x86_64 paging is already
+ * on (set up in boot.s) and vmm_init just records the PML4 it inherited
+ * so subsequent vmm_map calls can walk it. */
 void vmm_init(void);
 
 /* Install a mapping from virtual `virt` to physical `phys`, both 4 KiB
- * aligned.  Returns 0 on success, non-zero if the PDE is a 4 MiB PSE
- * entry or a PT allocation fails. */
-int vmm_map(uint32_t virt, uint32_t phys, uint32_t flags);
+ * aligned.  Returns 0 on success, non-zero if the relevant entry is
+ * already a large page or a page-table allocation fails. */
+int vmm_map(uintptr_t virt, uintptr_t phys, uint32_t flags);
 
-/* Physical address of the kernel page directory.  Used by the AP boot
- * trampoline (M18) so each AP can load CR3 before enabling paging. */
-uint32_t vmm_kernel_pd_phys(void);
+/* Physical address of the top-level page table.  Returned as uintptr_t
+ * so it works on both archs.  Used by the AP boot trampoline (M18) so
+ * each AP can load the same table into CR3 before enabling paging. */
+uintptr_t vmm_kernel_pd_phys(void);
 
-/* Install a single 4 MiB PSE mapping.  `virt` and `phys` must both be
- * 4 MiB aligned.  Returns 0 on success, non-zero if a non-PSE mapping
- * already exists in that PDE slot.  Useful for mapping MMIO-ish regions
- * (like a framebuffer) cheaply without burning a page table. */
-int vmm_map_4mib(uint32_t virt, uint32_t phys, uint32_t flags);
+/* Install a single "large page" mapping spanning 4 MiB of address space.
+ *
+ * Granule rationale:
+ *   - i386 PSE pages are exactly 4 MiB (one PDE covers 4 MiB).
+ *   - x86_64 long-mode large pages are 2 MiB at the PD level; the
+ *     impl installs TWO adjacent 2 MiB entries to keep the 4 MiB
+ *     contract.
+ *
+ * Both archs require `virt` and `phys` to be 4 MiB aligned.  Returns
+ * 0 on success, non-zero if a conflicting non-large mapping already
+ * exists.  Useful for cheap MMIO mappings (framebuffer, xHCI BARs)
+ * without burning a page table per 4 KiB. */
+int vmm_map_4mib(uintptr_t virt, uintptr_t phys, uint32_t flags);
 
 /* Remove a mapping.  Safe to call on unmapped addresses (no-op). */
-void vmm_unmap(uint32_t virt);
+void vmm_unmap(uintptr_t virt);
 
-/* Walk the tables and return the physical address `virt` maps to, with
- * the low 12 bits copied from `virt`.  Returns 0 if unmapped. */
-uint32_t vmm_translate(uint32_t virt);
+/* Walk the tables and return the physical address `virt` maps to,
+ * with the low 12 bits copied from `virt`.  Returns 0 if unmapped. */
+uintptr_t vmm_translate(uintptr_t virt);
 
 /* One-line diagnostics. */
 void vmm_print_status(void);
